Don't delete assets of a dist version if it is still latest.

+++ src/datapump/sync/sync.py
@@ -873,19 +873,22 @@ class DISTAlertsSync(Sync):
 
         jobs: List[Job] = []
 
-        # We are saving all old weekly dist alert rasters, but delete all the other
-        # assets in the next-to-last version that can be reproduced from the default
-        # raster, including the intensity raster, date_conf raster, COGs, and asset
-        # resampled to 10 meters, etc.
+        # We are saving all old weekly dist alert rasters, but delete in the
+        # next-to-last version all the other assets that can be reproduced from the
+        # default raster, including the intensity raster, date_conf raster, COGs, and
+        # asset resampled to 10 meters, etc.
         client = DataApiClient()
         dataset = client.get_dataset(self.dataset_name)
         next_to_last_version = sorted(dataset["versions"], reverse=True)[1]
-        assets = client.get_assets(self.dataset_name, next_to_last_version)
-        client.delete_uri_asset(assets, "/cog/intensity.tif")
-        client.delete_uri_asset(assets, "/cog/default.tif")
-        client.delete_uri_asset(assets, "/raster/epsg-4326/10/40000/intensity/")
-        client.delete_uri_asset(assets, "/date_conf/")
-        client.delete_uri_asset(assets, "/resample10/")
+
+        # Skip deletion if the next-to-last version is still marked "latest".
+        if next_to_last_version != latest_api_version:
+            assets = client.get_assets(self.dataset_name, next_to_last_version)
+            client.delete_uri_asset(assets, "/cog/intensity.tif")
+            client.delete_uri_asset(assets, "/cog/default.tif")
+            client.delete_uri_asset(assets, "/raster/epsg-4326/10/40000/intensity/")
+            client.delete_uri_asset(assets, "/date_conf/")
+            client.delete_uri_asset(assets, "/resample10/")
 
         slack_webhook("INFO", f"Starting dist-alerts jobs for {self.dataset_name}/{latest_release}")
